+++ packages/arch/src/cache/cache-options.ts
@@ -1,0 +1,6 @@
+import { Cache } from './cache'
+
+export interface CacheOptions {
+  ttl: number
+  cache: Cache
+}