+++ packages/arch/src/cache/invalidate-cache.ts
@@ -3,7 +3,7 @@ import { Class } from '@archimedes/utils'
 import { UseCase } from '../use-case/use-case'
 import { CacheInvalidations } from '../runner/cache-invalidations'
 
-export function EvictCache(clazz: Class): void {
+export function InvalidateCache(clazz: Class): void {
   const metadata: Class[] | undefined = Reflect.getMetadata('design:paramtypes', clazz)
   if (metadata !== undefined) {
     CacheInvalidations.set(