+++ packages/arch/src/runner/links/executor-link.spec.ts
@@ -5,26 +5,26 @@ import { UseCase } from '../../use-case/use-case'
 import { Link } from './link'
 
 describe('ExecutorLink', () => {
-  it('should execute', () => {
+  it('should execute', async () => {
     const { context, nextLink, executorLink } = setup()
     const useCase = mock<UseCase<unknown, unknown>>()
     when(context.useCase).thenReturn(instance(useCase))
     when(context.param).thenReturn(undefined)
     executorLink.setNext(instance(nextLink))
 
-    executorLink.next(instance(context))
+    await executorLink.next(instance(context))
 
     verify(useCase.internalExecute(anything())).once()
   })
 
-  it('should call next link', () => {
+  it('should call next link', async () => {
     const { context, nextLink, executorLink } = setup()
     const useCase = mock<UseCase<unknown, unknown>>()
     when(context.useCase).thenReturn(instance(useCase))
     when(context.param).thenReturn(undefined)
     executorLink.setNext(instance(nextLink))
 
-    executorLink.next(instance(context))
+    await executorLink.next(instance(context))
 
     verify(nextLink.next(anything())).once()
   })