+++ packages/components/stencil.config.ts
@@ -9,9 +9,6 @@ export const config: Config = {
       esmLoaderPath: '../loader'
     },
     {
-      type: 'dist-custom-elements-bundle'
-    },
-    {
       type: 'www',
       serviceWorker: null
     }