+++ packages/utils/src/http-client/http-client.spec.ts
@@ -1,4 +1,4 @@
-import { CreateOptions, HttpClient } from './http-client'
+import { HttpClientCreateOptions, HttpClient } from './http-client'
 
 describe('HttpClient', () => {
   it('should configure base url', async () => {
@@ -112,11 +112,32 @@ describe('HttpClient', () => {
       defaults: undefined
     })
   })
+
+  it('should handle empty responses', async () => {
+    const { httpClient } = setup({ result: '' })
+
+    const response = await httpClient.get('http://foo')
+
+    expect(response).toEqual({
+      headers: {},
+      options: {
+        method: 'GET'
+      },
+      result: '',
+      status: undefined
+    })
+  })
 })
 
-function setup<T>(options?: CreateOptions & Partial<{ result: T }>) {
+function setup<T>(options?: HttpClientCreateOptions & Partial<{ result: T }>) {
   const fetchMock = jest.fn()
-  fetchMock.mockImplementation(() => Promise.resolve({ json: () => Promise.resolve(options?.result), ok: true }))
+  fetchMock.mockImplementation(() =>
+    Promise.resolve({
+      text: () => Promise.resolve(JSON.stringify(options?.result ?? { foo: 'bar' })),
+      ok: true,
+      headers: new Headers()
+    })
+  )
   window.fetch = fetchMock
 
   return {