+++ packages/utils/src/http-client/http-client.ts
@@ -1,18 +1,29 @@
 import { HttpError } from './http-error'
 import { HttpParams } from './http-params'
+import { HttpStatusCode } from './http-status-code'
 
 type Url = string
 
-export type BeforeHook = (request: Request, options: Options) => void
-export type AfterHook = (response: Response & { result: unknown }, options: Options) => void
+export interface HttpClientResponse<Result> {
+  result: Result
+  status: HttpStatusCode
+  headers: HttpClientHeaders
+  options: RequestInit
+}
 
-export type Options = {
+export type HttpClientBeforeHook = (request: Request, options: HttpClientOptions) => void
+export type HttpClientAfterHook = (response: Response & { result: unknown }, options: HttpClientOptions) => void
+export type HttpClientHeaders = Record<string, string | number | boolean>
+
+export interface HttpClientOptions {
   baseUrl: Url
-  hooks: { before: BeforeHook[]; after: AfterHook[] }
+  hooks: { before: HttpClientBeforeHook[]; after: HttpClientAfterHook[] }
   defaults?: RequestInit
 }
 
-export type CreateOptions = Partial<Omit<Options, 'hooks'> & { hooks: { before?: BeforeHook[]; after?: AfterHook[] } }>
+export type HttpClientCreateOptions = Partial<
+  Omit<HttpClientOptions, 'hooks'> & { hooks: { before?: HttpClientBeforeHook[]; after?: HttpClientAfterHook[] } }
+>
 
 const defaultOptions = {
   baseUrl: '',
@@ -29,7 +40,7 @@ export class HttpClient {
     baseUrl = defaultOptions.baseUrl,
     hooks = { ...defaultOptions.hooks },
     defaults
-  }: CreateOptions = defaultOptions) {
+  }: HttpClientCreateOptions = defaultOptions) {
     const defaultBeforeHooks = hooks?.before ?? []
     const defaultAfterHooks = hooks?.after ?? []
 
@@ -40,31 +51,44 @@ export class HttpClient {
     })
   }
 
-  private constructor(private readonly options: Options) {}
+  private constructor(private readonly options: HttpClientOptions) {}
 
-  async get<Result>(url: Url, httpParams?: HttpParams): Promise<Result> {
+  async get<Result>(url: Url, httpParams?: HttpParams): Promise<HttpClientResponse<Result>> {
     return this.sendRequest(url, { method: 'GET' }, httpParams)
   }
 
-  async post<Result, Body>(url: string, body: Body, httpParams?: HttpParams): Promise<Result> {
+  async post<Body, Result = void>(
+    url: string,
+    body?: Body,
+    httpParams?: HttpParams
+  ): Promise<HttpClientResponse<Result>> {
     return this.sendRequest(url, { method: 'POST', body: this.getParsedBody(body) }, httpParams)
   }
 
-  async put<Result, Body>(url: string, body: Body, httpParams?: HttpParams): Promise<Result> {
+  async put<Body, Result = void>(
+    url: string,
+    body?: Body,
+    httpParams?: HttpParams
+  ): Promise<HttpClientResponse<Result>> {
     return this.sendRequest(url, { method: 'PUT', body: this.getParsedBody(body) }, httpParams)
   }
 
-  async delete<Result>(url: string, httpParams?: HttpParams): Promise<Result> {
+  async delete<Result = void>(url: string, httpParams?: HttpParams): Promise<HttpClientResponse<Result>> {
     return this.sendRequest(url, { method: 'DELETE' }, httpParams)
   }
 
-  private async sendRequest<Result>(url: string, options: RequestInit, httpParams?: HttpParams): Promise<Result> {
+  private async sendRequest<Result>(
+    url: string,
+    options: RequestInit,
+    httpParams?: HttpParams
+  ): Promise<HttpClientResponse<Result>> {
     const request = this.getRequest(url, httpParams)
     this.options.hooks.before.forEach(hook => hook(request, this.options))
     const response = await fetch(request, { ...this.options.defaults, ...options })
-    const result = await response.json()
+    const resultAsString = await response.text()
+    const result = resultAsString.length !== 0 ? JSON.parse(resultAsString) : undefined
     this.options.hooks.after.forEach(hook => hook({ ...response, result }, this.options))
-    return this.getResponse(response)
+    return this.getResponse(response, result, options)
   }
 
   private getRequest(url: string, httpParams: HttpParams | undefined) {
@@ -84,11 +108,21 @@ export class HttpClient {
     return JSON.stringify(body)
   }
 
-  private async getResponse<Result>(response: Response) {
+  private async getResponse<Result>(
+    response: Response,
+    result: Result,
+    options: RequestInit
+  ): Promise<HttpClientResponse<Result>> {
     if (!response.ok) {
       throw new HttpError({ name: response.status.toString(), message: response.statusText })
     }
-    const json = await response.json()
-    return json as Result
+
+    const headers: HttpClientHeaders = {}
+
+    response.headers.forEach((value, key) => {
+      headers[key] = value
+    })
+
+    return { result: result as Result, headers, options, status: response.status }
   }
 }