+++ packages/utils/CHANGELOG.md
@@ -1,3 +1,73 @@
+# @archimedes/utils [2.0.0-beta.2](https://github.com/archimedes-projects/archimedes-js/compare/@archimedes/utils@2.0.0-beta.1...@archimedes/utils@2.0.0-beta.2) (2022-01-10)
+
+### Features
+
+-   make body optional in put and post ([1674c2a](https://github.com/archimedes-projects/archimedes-js/commit/1674c2a0e8a19f2583b50bfb64f5cd2af8b85198))
+
+# @archimedes/utils [2.0.0-beta.1](https://github.com/archimedes-projects/archimedes-js/compare/@archimedes/utils@1.0.0...@archimedes/utils@2.0.0-beta.1) (2022-01-05)
+
+### Features
+
+-   add default type values ([9ffcb96](https://github.com/archimedes-projects/archimedes-js/commit/9ffcb96bdc4e8d6e54eb6a98b3cebd7e69a8cf4d))
+-   add documentation ([b7667d7](https://github.com/archimedes-projects/archimedes-js/commit/b7667d7b78942a27ba4316300a14057a499d7c1b))
+-   add links to the official documentation ([decbebb](https://github.com/archimedes-projects/archimedes-js/commit/decbebbf7d5b41896819d86d896293c165b54021))
+-   add options ([cddebdb](https://github.com/archimedes-projects/archimedes-js/commit/cddebdb4341a07baafc5bfaca6e61cddb8044526))
+-   add parcel 2 configuration ([c6a3d61](https://github.com/archimedes-projects/archimedes-js/commit/c6a3d6183a15affe71eed9e74b8bed380a031e90))
+-   fix typo ([1446e66](https://github.com/archimedes-projects/archimedes-js/commit/1446e666071dfb95f74a7f98ad87eca07b7652af))
+-   force breaking change ([eaa774e](https://github.com/archimedes-projects/archimedes-js/commit/eaa774ea427be406071db5ddf544eabc87064c89))
+-   force breaking change ([2f50d1b](https://github.com/archimedes-projects/archimedes-js/commit/2f50d1b29b765378610d5b978bb1fc11a174453c))
+-   use export type ([107c758](https://github.com/archimedes-projects/archimedes-js/commit/107c758c510edebfff4efe44c5e95fe26b9c7d55))
+
+-   feat!: force breaking change ([980a4a1](https://github.com/archimedes-projects/archimedes-js/commit/980a4a1a672221df950dc41b820a0fb8e3ae47d5))
+
+### BREAKING CHANGES
+
+-   force breaking change
+-   force breaking change
+-   force breaking change
+
+# @archimedes/utils [1.0.0-beta.2](https://github.com/archimedes-projects/archimedes-js/compare/@archimedes/utils@1.0.0-beta.1...@archimedes/utils@1.0.0-beta.2) (2022-01-05)
+
+### Features
+
+-   force breaking change ([2f50d1b](https://github.com/archimedes-projects/archimedes-js/commit/2f50d1b29b765378610d5b978bb1fc11a174453c))
+
+### BREAKING CHANGES
+
+-   force breaking change
+
+# @archimedes/utils 1.0.0-beta.1 (2022-01-04)
+
+### Features
+
+-   add "as" method to duration to get a given duration in a time unit ([44d6786](https://github.com/archimedes-projects/archimedes-js/commit/44d67860caefc0c04f0f74b14ebedcbd2e9a202e))
+-   add default headers ([1ea6510](https://github.com/archimedes-projects/archimedes-js/commit/1ea651087c8259cbc889b4b6e54d97881d854cfd))
+-   add default type values ([9ffcb96](https://github.com/archimedes-projects/archimedes-js/commit/9ffcb96bdc4e8d6e54eb6a98b3cebd7e69a8cf4d))
+-   add difference method to get the difference between two dates ([a42dcc3](https://github.com/archimedes-projects/archimedes-js/commit/a42dcc3094e4d80f28aa445514b0c96fa52028c5))
+-   add documentation ([b7667d7](https://github.com/archimedes-projects/archimedes-js/commit/b7667d7b78942a27ba4316300a14057a499d7c1b))
+-   add links to the official documentation ([decbebb](https://github.com/archimedes-projects/archimedes-js/commit/decbebbf7d5b41896819d86d896293c165b54021))
+-   add options ([cddebdb](https://github.com/archimedes-projects/archimedes-js/commit/cddebdb4341a07baafc5bfaca6e61cddb8044526))
+-   add parcel 2 configuration ([c6a3d61](https://github.com/archimedes-projects/archimedes-js/commit/c6a3d6183a15affe71eed9e74b8bed380a031e90))
+-   create range function ([0c980b2](https://github.com/archimedes-projects/archimedes-js/commit/0c980b29b22b3165c148aaaa6a2bff031f2abc4d))
+-   fix typo ([1446e66](https://github.com/archimedes-projects/archimedes-js/commit/1446e666071dfb95f74a7f98ad87eca07b7652af))
+-   return result in the after hook ([4f29d99](https://github.com/archimedes-projects/archimedes-js/commit/4f29d99cb2c4916f31975230e8c98280ecd7d993))
+-   use bili and yarn to build ([0fdbccc](https://github.com/archimedes-projects/archimedes-js/commit/0fdbcccff1bb1704a1579531c798a1d398218a50))
+-   use export type ([107c758](https://github.com/archimedes-projects/archimedes-js/commit/107c758c510edebfff4efe44c5e95fe26b9c7d55))
+-   use request defaults ([901210e](https://github.com/archimedes-projects/archimedes-js/commit/901210e5fa3bf505c20da597302a90aec8daf02b))
+-   **arch:** add base url ([f2a2f0b](https://github.com/archimedes-projects/archimedes-js/commit/f2a2f0b55e0d56b498b93f411d203060ade9dfe9))
+
+-   feat!: force breaking change ([980a4a1](https://github.com/archimedes-projects/archimedes-js/commit/980a4a1a672221df950dc41b820a0fb8e3ae47d5))
+-   chore!: release version 1.0 ([2e189e4](https://github.com/archimedes-projects/archimedes-js/commit/2e189e4f24c216edb0cf1706003242d115bc0e64))
+
+### Bug Fixes
+
+-   remove window ([8aba18a](https://github.com/archimedes-projects/archimedes-js/commit/8aba18abfd1a864b9cae93f54935fe192da5acf1))
+
+### BREAKING CHANGES
+
+-   force breaking change
+-   release new version
+
 # Change Log
 
 All notable changes to this project will be documented in this file.
@@ -5,14 +75,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.5.0](https://github.com/archimedes-projects/archimedes-js/compare/v1.4.0...v1.5.0) (2021-11-23)
 
-
 ### Features
 
-* return result in the after hook ([4f29d99](https://github.com/archimedes-projects/archimedes-js/commit/4f29d99cb2c4916f31975230e8c98280ecd7d993))
-
-
-
-
+-   return result in the after hook ([4f29d99](https://github.com/archimedes-projects/archimedes-js/commit/4f29d99cb2c4916f31975230e8c98280ecd7d993))
 
 # [1.4.0](https://github.com/archimedes-projects/archimedes-js/compare/v1.4.0-beta.1...v1.4.0) (2021-09-16)